📝 Document functions and remove install path config

+++ admin/index.php
@@ -1,64 +1,89 @@
 <?php
 
+// Get config from json file
 $GLOBALS["config"] = json_decode(file_get_contents(__DIR__ . "/../config.json"), true);
 
 function add()
 {
+	// Skip if not submitted or action is not add
 	if (!isset($_POST["submit"]) || $_POST["action"] != "add") return "&nbsp;";
+
+	// Get uploaded file and check for file errors
 	$file = $_FILES["file"];
 	if ($file["error"] != 0) return "Fehler beim Hochladen der Datei";
 	if (file_exists(__DIR__ . "/../news/" . $file["name"])) return "Datei existiert bereits";
 	if (!getimagesize($file["tmp_name"])) return "Datei ist kein Bild";
+	if ($file["size"] > 50 * 1024 * 1024) return "Datei ist zu groß";
+
+	// Store file
 	$move = move_uploaded_file($file["tmp_name"], __DIR__ . "/../news/" . $file["name"]);
 	if (!$move) return "Fehler beim Speichern der Datei";
+
+	// Add new news to config
 	array_push($GLOBALS["config"]["news"], $file["name"]);
 	$configUpdated = file_put_contents(__DIR__ . "/../config.json", json_encode($GLOBALS["config"]));
 	if (!$configUpdated) {
 		unlink(__DIR__ . "/../news/" . $file["name"]);
 		return "Fehler beim Speichern der Konfiguration";
 	}
+
 	return "Datei erfolgreich hochgeladen";
 }
 
 function newsSelect()
 {
+	// Output select option for each news
 	foreach ($GLOBALS["config"]["news"] as $news) {
 		echo "<option value='$news'>$news</option>";
 	}
 }
 
-function delete()
+function remove()
 {
+	// Skip if not submitted or action is not remove
 	if (!isset($_POST["submit"]) || $_POST["action"] != "remove") return "&nbsp;";
+
+	// Get news from POST
 	$news = $_POST["news"];
 	if (!$news) return "Bitte eine News auswählen";
+
+	// Check if news exists
 	$index = array_search($news, $GLOBALS["config"]["news"]);
 	if (!$index) return "Datei nicht gefunden";
+
+	// Delete news file
 	$delete = unlink(__DIR__ . "/../news/" . $news);
 	if (!$delete) return "Fehler beim Löschen der Datei";
+
+	// Remove news from config
 	array_splice($GLOBALS["config"]["news"], $index, 1);
 	$configUpdated = file_put_contents(__DIR__ . "/../config.json", json_encode($GLOBALS["config"]));
 	if (!$configUpdated) return "Fehler beim Speichern der Konfiguration";
+
 	return "Datei erfolgreich gelöscht";
 }
 
 function settings()
 {
+	// Skip if not submitted or action is not settings
 	if (!isset($_POST["submit"]) || $_POST["action"] != "settings") return "&nbsp;";
+
+	// Get settings from POST and validate them
 	$departureUrl = $_POST["departureUrl"];
 	$newsInterval = $_POST["newsInterval"];
 	$departureInterval = $_POST["departureInterval"];
 	$reloadInterval = $_POST["reloadInterval"];
-	$installPath = $_POST["installPath"];
 	if (!$departureUrl) return "Bitte eine URL eingeben";
 	if (!$newsInterval || !$departureInterval || !$reloadInterval) return "Bitte eine Zahl eingeben";
+
+	// Update config with new settings
 	$GLOBALS["config"]["departureUrl"] = $departureUrl;
 	$GLOBALS["config"]["newsInterval"] = $newsInterval;
 	$GLOBALS["config"]["departureInterval"] = $departureInterval;
 	$GLOBALS["config"]["reloadInterval"] = $reloadInterval;
-	$GLOBALS["config"]["installPath"] = $installPath;
 	$configUpdated = file_put_contents(__DIR__ . "/../config.json", json_encode($GLOBALS["config"]));
 	if (!$configUpdated) return "Fehler beim Speichern der Konfiguration";
+
 	return "Einstellungen erfolgreich geändert";
 }
 
@@ -100,7 +125,7 @@ function settings()
 				<?php newsSelect(); ?>
 			</select>
 			<input type="submit" name="submit" value="Löschen">
-			<label><?php echo delete(); ?></label>
+			<label><?php echo remove(); ?></label>
 		</form>
 		<br>
 		<h3>Einstellungen</h3>
@@ -111,7 +136,6 @@ function settings()
 			<label for="newsInterval">News Intervall (in Sekunden)</label> <input type="number" name="newsInterval" value="<?php echo $GLOBALS["config"]["newsInterval"]; ?>"><br>
 			<label for="departureInterval">Abfahrten Intervall (in Sekunden)</label> <input type="number" name="departureInterval" value="<?php echo $GLOBALS["config"]["departureInterval"]; ?>"><br>
 			<label for="reloadInterval">Neulade Intervall (in Sekunden)</label> <input type="number" name="reloadInterval" value="<?php echo $GLOBALS["config"]["reloadInterval"]; ?>"><br>
-			<label for="installPath">Installation Verzeichnis</label> <input type="text" name="installPath" value="<?php echo $GLOBALS["config"]["installPath"]; ?>"><br>
 			<input type="submit" name="submit" value="Ändern">
 			<label><?php echo $settings; ?></label>
 		</form>
@@ -122,11 +146,6 @@ function settings()
 			Version 1.0<br>
 			Entwickelt von <a href="https://friedinger.org/">Manuel Weis</a>
 		</p>
-		<script>
-			document.querySelectorAll("a").forEach(element => {
-				element.setAttribute("href", element.getAttribute("href") + window.location.search);
-			});
-		</script>
 	</main>
 </body>
